+++ src/components/settings/SettingsFieldButton.jsx
@@ -3,6 +3,7 @@
  * @param {string} [props.className]
  * @param {import('react').MouseEventHandler<HTMLButtonElement>} [props.onClick]
  * @param {boolean} [props.danger]
+ * @param {boolean} [props.inverted]
  * @param {import('react').FC<any>} props.Icon
  * @param {import('react').ReactNode} props.children
  * @param {boolean} [props.disabled]
@@ -13,6 +14,7 @@ export default function SettingsFieldButton({
   Icon,
   children,
   danger,
+  inverted,
   disabled = !onClick,
 }) {
   return (
@@ -20,9 +22,13 @@ export default function SettingsFieldButton({
       className={
         'w-full flex flex-row items-center outline rounded p-2 my-4' +
         ' ' +
-        'disabled:opacity-30 enabled:hover:text-white' +
+        'disabled:opacity-30' +
         ' ' +
-        (danger ? 'enabled:hover:bg-red-500' : 'enabled:hover:bg-black') +
+        (danger
+          ? 'enabled:hover:bg-red-500 enabled:hover:text-white'
+          : inverted
+            ? 'enabled:hover:bg-white enabled:hover:text-black bg-black text-white'
+            : 'enabled:hover:bg-black enabled:hover:text-white') +
         ' ' +
         className
       }